refactor(backend): move into events
