+++ packages/backend/convex/_generated/api.d.ts
@@ -10,7 +10,6 @@
 
 import type * as auth from "../auth.js";
 import type * as data_ml_eventRec from "../data_ml/eventRec.js";
-import type * as data_ml_events from "../data_ml/events.js";
 import type * as data_ml_friendRecs from "../data_ml/friendRecs.js";
 import type * as data_ml_friends from "../data_ml/friends.js";
 import type * as data_ml_universal from "../data_ml/universal.js";
@@ -20,7 +19,6 @@ import type * as eventSeedsStatic from "../eventSeedsStatic.js";
 import type * as events from "../events.js";
 import type * as eventsIngest from "../eventsIngest.js";
 import type * as files from "../files.js";
-import type * as posts from "../posts.js";
 import type * as seed from "../seed.js";
 import type * as users from "../users.js";
 
@@ -33,7 +31,6 @@ import type {
 declare const fullApi: ApiFromModules<{
   auth: typeof auth;
   "data_ml/eventRec": typeof data_ml_eventRec;
-  "data_ml/events": typeof data_ml_events;
   "data_ml/friendRecs": typeof data_ml_friendRecs;
   "data_ml/friends": typeof data_ml_friends;
   "data_ml/universal": typeof data_ml_universal;
@@ -43,7 +40,6 @@ declare const fullApi: ApiFromModules<{
   events: typeof events;
   eventsIngest: typeof eventsIngest;
   files: typeof files;
-  posts: typeof posts;
   seed: typeof seed;
   users: typeof users;
 }>;
